% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{droughts}
\alias{droughts}
\title{Periods Between Rain Events}
\format{
A data frame with 2 columns:
\describe{
\item{rownames}{instance of each entry,}
\item{length}{the length of time from the completion of the last rain event to the beginning of the next rain event,}
\item{year}{the calendar year.}
}
}
\usage{
droughts
}
\description{
Data collected at Winnipeg International Airport (Canada) on periods (in days) between rain events.
}
\examples{
\dontrun{
  boxplot(length ~ year, data = droughts)
  boxplot(log(length) ~ year, data = droughts)
  hist(droughts$length, main = "Winnipeg Droughts", xlab = "length (in days)")
  hist(log(droughts$length), main = "Winnipeg Droughts", xlab = "length (in days, log scale)")
}
}
\keyword{datasets}
