% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNegBin.R
\name{runNegBin}
\alias{runNegBin}
\title{Negative Binomial App}
\usage{
runNegBin()
}
\description{
This function allows the Negative Binomial shiny app to run. The app gives demonstrations
of Negative Binomial distributions as well as some interactivity with different regression
that the model can predict. It shows also provides differences between the Negative
Binomial regression and other common regressive techniques.
}
