% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bikes_bridges}
\alias{bikes_bridges}
\title{Bicycle Counts for East River Bridges}
\format{
A data frame containing 215 observations on 2 date variables, 2 temperature ranges, precipitation amounts, 4 different bridges and a total number of bikes.
\describe{
\item{date}{Factor with each date,}
\item{day}{factor with the day,}
\item{temp_high}{number with high temperature recorded each day,}
\item{temp_low}{number with low temperature recorded each day,}
\item{precipitation}{precipitation percentages for each day,}
\item{Brooklyn_bridge}{number of bikes that crossed the Brooklyn bridge,}
\item{Manhattan_bridge}{number of bikes that crossed the Manhattan bridge,}
\item{Williamsburg_bridge}{number of bikes that crossed the Williamsburg bridge,}
\item{Queensboro_bridge}{number of bikes that crossed the Queensboro bridge,}
\item{total}{total number of bikes that crossed every bridge for the day,}
}
}
\source{
NYCDOT Bicycle Counts 2017.

https://data.cityofnewyork.us/Transportation/Bicycle-Counts-for-East-River-Bridges-Historical-/gua4-p9wg/about_data
}
\usage{
bikes_bridges
}
\description{
Daily total of bike counts conducted monthly on the Brooklyn Bridge, Manhattan Bridge, Williamsburg Bridge, and Queensboro Bridge.
}
\details{
The data are from the New York Department of Transportation for 2017
}
\keyword{datasets}
